File picker

+++ app/src/main/java/com/example/scoped_storage_example/media_store/data/models/MediaFile.kt
@@ -6,6 +6,6 @@ data class MediaFile(
     val uri: Uri?,
     val name: String,
     val type: String,
-    val sizeKb: Int,
+    val sizeKb: Long,
     val date: Long
 )
